+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All Notable changes to `http-transitions` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [v0.4.0] - 2018-10-26
+
+### Added
+- extracted `TransitionFactory` interface
+- updated Laravel service provider to use `LaravelTransitionFactory`
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## [v0.2.0] - 2017-09-18
 
 ### Added